move language settings dialog to separate file
